Se crea el servicio rest y se realiza la implementación de la lógica de negocio, tambien se define la capa de aplicación encargada de intermediar entre el servicio rest y la lógica de negocio

+++ src/main/java/com/yesid/mudanza/domain/model/MovingElement.java
@@ -2,13 +2,13 @@ package com.yesid.mudanza.domain.model;
 
 public class MovingElement {
 
-	private Double weight;
+	private int weight;
 	
-	public MovingElement(Double weight) {
+	public MovingElement(int weight) {
 		this.weight = weight;
 	}
 
-	public Double getWeight() {
+	public int getWeight() {
 		return weight;
 	}	
 }
